+++ scripts/aggregate-translations/aggregate-translations.js
@@ -19,22 +19,19 @@ const customDirectories = (baseDirectory, directories) =>
 
 const isFile = filePath => (fse.existsSync(filePath) && !fse.lstatSync(filePath).isDirectory());
 
-const configFile = (configPath) => {
-  if (configPath) {
+const loadConfigFile = (configPath) => {
+  const i18nConfigPath = path.resolve(process.cwd(), configPath);
+  if (isFile(i18nConfigPath)) {
     // eslint-disable-next-line global-require, import/no-dynamic-require
-    return require(configPath);
-  }
-
-  const localPath = path.resolve(process.cwd(), 'terraI18n.config.js');
-  if (isFile(localPath)) {
-    // eslint-disable-next-line global-require, import/no-dynamic-require
-    return require(localPath);
+    return require(i18nConfigPath);
   }
   return {};
 };
 
 const defaults = (options = {}) => {
-  const config = configFile(options.configPath);
+  const configPath = options.configPath || './terraI18n.config.js';
+  const config = loadConfigFile(configPath);
+
   const defaultConfig = {
     baseDir: options.baseDir || config.baseDir || process.cwd(),
     directories: options.directories || config.directories || [],